# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthLoginInput {
  /** @minLength 1 */
  token: string;
}

export interface AuthUser {
  sub: string;
  /** @nullable */
  email?: string | null;
  roles: string[];
}

export interface HealthStatus {
  status: string;
}

export type DashboardScanStatus = typeof DashboardScanStatus[keyof typeof DashboardScanStatus];


export const DashboardScanStatus = {
  monitoring: 'monitoring',
  scanning: 'scanning',
  paused: 'paused',
} as const;

export type DashboardFindingsBySeverity = {
  critical: number;
  high: number;
  medium: number;
  low: number;
};

export interface Dashboard {
  complianceScore: number;
  openFindings: number;
  criticalFindings: number;
  protectedRecords: number;
  monitoredSources: number;
  lastScanAt: string;
  scanStatus: DashboardScanStatus;
  findingsBySeverity: DashboardFindingsBySeverity;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  finding: 'finding',
  scan: 'scan',
  masking: 'masking',
  report: 'report',
  system: 'system',
} as const;

/**
 * @nullable
 */
export type ActivitySeverity = typeof ActivitySeverity[keyof typeof ActivitySeverity] | null;


export const ActivitySeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  title: string;
  description: string;
  createdAt: string;
  /** @nullable */
  severity?: ActivitySeverity;
}

export type FindingDataType = typeof FindingDataType[keyof typeof FindingDataType];


export const FindingDataType = {
  email: 'email',
  phone: 'phone',
  national_id: 'national_id',
  credit_card: 'credit_card',
  health: 'health',
  address: 'address',
  password: 'password',
} as const;

export type FindingSeverity = typeof FindingSeverity[keyof typeof FindingSeverity];


export const FindingSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


export const FindingStatus = {
  open: 'open',
  in_review: 'in_review',
  resolved: 'resolved',
} as const;

export interface Finding {
  id: string;
  title: string;
  dataType: FindingDataType;
  source: string;
  location: string;
  severity: FindingSeverity;
  status: FindingStatus;
  records: number;
  detectedAt: string;
  regulation: string;
  recommendation: string;
  /** @nullable */
  sample?: string | null;
}

export type FindingUpdateStatus = typeof FindingUpdateStatus[keyof typeof FindingUpdateStatus];


export const FindingUpdateStatus = {
  open: 'open',
  in_review: 'in_review',
  resolved: 'resolved',
} as const;

export interface FindingUpdate {
  status: FindingUpdateStatus;
}

export type DataSourceKind = typeof DataSourceKind[keyof typeof DataSourceKind];


export const DataSourceKind = {
  postgresql: 'postgresql',
  mysql: 'mysql',
  mongodb: 'mongodb',
  snowflake: 'snowflake',
  bigquery: 'bigquery',
} as const;

export type DataSourceEnvironment = typeof DataSourceEnvironment[keyof typeof DataSourceEnvironment];


export const DataSourceEnvironment = {
  production: 'production',
  staging: 'staging',
  development: 'development',
} as const;

export type DataSourceStatus = typeof DataSourceStatus[keyof typeof DataSourceStatus];


export const DataSourceStatus = {
  healthy: 'healthy',
  warning: 'warning',
  offline: 'offline',
} as const;

export interface DataSource {
  id: string;
  name: string;
  kind: DataSourceKind;
  environment: DataSourceEnvironment;
  status: DataSourceStatus;
  lastScanAt: string;
  tables: number;
  records: number;
  findings: number;
}

export interface Rule {
  id: string;
  name: string;
  category: string;
  regulation: string;
  enabled: boolean;
  detections: number;
  lastTriggered: string;
}

export interface ScanInput {
  sourceId: string;
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
} as const;

export interface Scan {
  id: string;
  sourceId: string;
  status: ScanStatus;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  findingsCreated?: number;
}

export type ReportInputPeriod = typeof ReportInputPeriod[keyof typeof ReportInputPeriod];


export const ReportInputPeriod = {
  last_24h: 'last_24h',
  last_7d: 'last_7d',
  last_30d: 'last_30d',
  quarter: 'quarter',
} as const;

export interface ReportInput {
  name: string;
  period: ReportInputPeriod;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  generating: 'generating',
  ready: 'ready',
} as const;

export type ReportFormat = typeof ReportFormat[keyof typeof ReportFormat];


export const ReportFormat = {
  pdf: 'pdf',
  csv: 'csv',
} as const;

export interface Report {
  id: string;
  name: string;
  period: string;
  status: ReportStatus;
  createdAt: string;
  findings: number;
  complianceScore: number;
  format?: ReportFormat;
}

export interface MaskingInput {
  sourceId: string;
  fields: string[];
}

export type MaskingPreviewRowsItem = {[key: string]: string};

export interface MaskingPreview {
  sourceId: string;
  records: number;
  maskedFields: string[];
  rows: MaskingPreviewRowsItem[];
}

export type ListFindingsParams = {
status?: ListFindingsStatus;
severity?: ListFindingsSeverity;
};

export type ListFindingsStatus = typeof ListFindingsStatus[keyof typeof ListFindingsStatus];


export const ListFindingsStatus = {
  open: 'open',
  in_review: 'in_review',
  resolved: 'resolved',
} as const;

export type ListFindingsSeverity = typeof ListFindingsSeverity[keyof typeof ListFindingsSeverity];


export const ListFindingsSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;