# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and session
  - name: dashboard
    description: Overview and activity
  - name: findings
    description: Sensitive data findings
  - name: sources
    description: Connected data sources
  - name: rules
    description: Detection rules
  - name: reports
    description: Audit reports
  - name: masking
    description: Test data anonymization
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      security: []
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login with bootstrap token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Authenticated (session cookie set)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Missing bootstrap token in request body
        "401":
          description: Invalid bootstrap token
  /auth/logout:
    post:
      operationId: authLogout
      tags: [auth]
      summary: Logout (clears session cookie)
      security: []
      responses:
        "204":
          description: Session cleared
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get authenticated user identity
      responses:
        "200":
          description: Authenticated user identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get compliance dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Get recent compliance activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /findings:
    get:
      operationId: listFindings
      tags: [findings]
      summary: List privacy findings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, in_review, resolved]
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [critical, high, medium, low]
      responses:
        "200":
          description: Findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
  /findings/{id}:
    patch:
      operationId: updateFinding
      tags: [findings]
      summary: Update a finding status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingUpdate"
      responses:
        "200":
          description: Updated finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          description: Finding not found
  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List monitored data sources
      responses:
        "200":
          description: Data sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataSource"
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List detection rules
      responses:
        "200":
          description: Detection rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
  /scans:
    post:
      operationId: startScan
      tags: [sources]
      summary: Start a privacy scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "202":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List audit reports
      responses:
        "200":
          description: Audit reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Generate an audit report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /masking/preview:
    post:
      operationId: previewMasking
      tags: [masking]
      summary: Preview anonymization for test data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaskingInput"
      responses:
        "200":
          description: Masked data preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaskingPreview"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthLoginInput:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 1 }
    AuthUser:
      type: object
      required: [sub, roles]
      properties:
        sub: { type: string }
        email: { type: ["string", "null"] }
        roles:
          type: array
          items: { type: string }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Dashboard:
      type: object
      required: [complianceScore, openFindings, criticalFindings, protectedRecords, monitoredSources, lastScanAt, scanStatus, findingsBySeverity]
      properties:
        complianceScore: { type: number }
        openFindings: { type: number }
        criticalFindings: { type: number }
        protectedRecords: { type: number }
        monitoredSources: { type: number }
        lastScanAt: { type: string }
        scanStatus: { type: string, enum: [monitoring, scanning, paused] }
        findingsBySeverity:
          type: object
          required: [critical, high, medium, low]
          properties:
            critical: { type: number }
            high: { type: number }
            medium: { type: number }
            low: { type: number }
    Activity:
      type: object
      required: [id, type, title, description, createdAt]
      properties:
        id: { type: string }
        type: { type: string, enum: [finding, scan, masking, report, system] }
        title: { type: string }
        description: { type: string }
        createdAt: { type: string }
        severity: { type: ["string", "null"], enum: [critical, high, medium, low, null] }
    Finding:
      type: object
      required: [id, title, dataType, source, location, severity, status, records, detectedAt, regulation, recommendation]
      properties:
        id: { type: string }
        title: { type: string }
        dataType: { type: string, enum: [email, phone, national_id, credit_card, health, address, password] }
        source: { type: string }
        location: { type: string }
        severity: { type: string, enum: [critical, high, medium, low] }
        status: { type: string, enum: [open, in_review, resolved] }
        records: { type: number }
        detectedAt: { type: string }
        regulation: { type: string }
        recommendation: { type: string }
        sample: { type: ["string", "null"] }
    FindingUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, in_review, resolved] }
    DataSource:
      type: object
      required: [id, name, kind, environment, status, lastScanAt, tables, records, findings]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [postgresql, mysql, mongodb, snowflake, bigquery] }
        environment: { type: string, enum: [production, staging, development] }
        status: { type: string, enum: [healthy, warning, offline] }
        lastScanAt: { type: string }
        tables: { type: number }
        records: { type: number }
        findings: { type: number }
    Rule:
      type: object
      required: [id, name, category, regulation, enabled, detections, lastTriggered]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        regulation: { type: string }
        enabled: { type: boolean }
        detections: { type: number }
        lastTriggered: { type: string }
    ScanInput:
      type: object
      required: [sourceId]
      properties:
        sourceId: { type: string }
    Scan:
      type: object
      required: [id, sourceId, status, startedAt]
      properties:
        id: { type: string }
        sourceId: { type: string }
        status: { type: string, enum: [queued, running, completed] }
        startedAt: { type: string }
        completedAt: { type: ["string", "null"] }
        findingsCreated: { type: number }
    ReportInput:
      type: object
      required: [name, period]
      properties:
        name: { type: string }
        period: { type: string, enum: [last_24h, last_7d, last_30d, quarter] }
    Report:
      type: object
      required: [id, name, period, status, createdAt, findings, complianceScore]
      properties:
        id: { type: string }
        name: { type: string }
        period: { type: string }
        status: { type: string, enum: [generating, ready] }
        createdAt: { type: string }
        findings: { type: number }
        complianceScore: { type: number }
        format: { type: string, enum: [pdf, csv] }
    MaskingInput:
      type: object
      required: [sourceId, fields]
      properties:
        sourceId: { type: string }
        fields:
          type: array
          items: { type: string }
    MaskingPreview:
      type: object
      required: [sourceId, records, maskedFields, rows]
      properties:
        sourceId: { type: string }
        records: { type: number }
        maskedFields:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            type: object
            additionalProperties:
              type: string