# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  sub: string;
  /** @nullable */
  email?: string | null;
  roles: string[];
}